ai not on drifter

+++ links.zsh
@@ -60,7 +60,11 @@ ln -sf $DIR/environment/environment.d/10-common.conf $XDG_CONFIG_HOME/environmen
 
 CACHE=/run/media/$USER/data/.cache
 
-[[ -d $CACHE/llama.cpp ]] || mkdir "$CACHE"/llama.cpp
-[[ -e $XDG_CACHE_HOME/llama.cpp ]] && rm -rf "$XDG_CACHE_HOME"/llama.cpp
-ln -s "$CACHE"/llama.cpp "$XDG_CACHE_HOME"/llama.cpp
+if [[ $HOST =~ ^(player|worker)$ ]]; then
+
+  [[ -d $CACHE/llama.cpp ]] || mkdir "$CACHE"/llama.cpp
+  [[ -e $XDG_CACHE_HOME/llama.cpp ]] && rm -rf "$XDG_CACHE_HOME"/llama.cpp
+  ln -s "$CACHE"/llama.cpp "$XDG_CACHE_HOME"/llama.cpp
+
+fi
